+++ src/Oro/Component/ChainProcessor/Tests/Unit/DependencyInjection/LoadProcessorsCompilerPassTest.php
@@ -4,6 +4,7 @@ namespace Oro\Component\ChainProcessor\Tests\Unit\DependencyInjection;
 
 use Oro\Component\ChainProcessor\DependencyInjection\LoadProcessorsCompilerPass;
 use Oro\Component\ChainProcessor\ProcessorBagConfigBuilder;
+use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 
@@ -268,7 +269,7 @@ class LoadProcessorsCompilerPassTest extends \PHPUnit\Framework\TestCase
         );
     }
 
-    public function testProcessForNotPublicProcessor()
+    public function testProcessForDecoratedProcessors()
     {
         $container = new ContainerBuilder();
         $container->setParameter('kernel.debug', false);
@@ -277,12 +278,34 @@ class LoadProcessorsCompilerPassTest extends \PHPUnit\Framework\TestCase
         $processorBagConfigBuilder = new Definition(ProcessorBagConfigBuilder::class, [$groups]);
 
         $processor1 = new Definition('Test\Processor1');
-        $processor1->setPublic(false);
         $processor1->addTag('processor');
+        $processor1Decorator = new ChildDefinition('parent');
+        $processor1Decorator->setDecoratedService('processor1');
+
+        $processor2 = new Definition('Test\Processor2');
+        $processor2->addTag('processor');
+        $processor2Decorator1 = new ChildDefinition('parent');
+        $processor2Decorator1->setDecoratedService('processor2');
+        $processor2Decorator2 = new ChildDefinition('parent');
+        $processor2Decorator2->setDecoratedService('processor2');
+
+        $processor3 = new Definition('Test\Processor3');
+        $processor3->addTag('processor');
+        $processor3Decorator1 = new ChildDefinition('parent');
+        $processor3Decorator1->setDecoratedService('processor3', null, -10);
+        $processor3Decorator2 = new ChildDefinition('parent');
+        $processor3Decorator2->setDecoratedService('processor3');
 
         $container->addDefinitions([
             'processor_bag_config_builder' => $processorBagConfigBuilder,
-            'processor1'                   => $processor1
+            'processor1'                   => $processor1,
+            'processor1.decorator'         => $processor1Decorator,
+            'processor2'                   => $processor2,
+            'processor2.decorator1'        => $processor2Decorator1,
+            'processor2.decorator2'        => $processor2Decorator2,
+            'processor3'                   => $processor3,
+            'processor3.decorator1'        => $processor3Decorator1,
+            'processor3.decorator2'        => $processor3Decorator2
         ]);
 
         $compilerPass = new LoadProcessorsCompilerPass('processor_bag_config_builder', 'processor');
@@ -294,9 +317,16 @@ class LoadProcessorsCompilerPassTest extends \PHPUnit\Framework\TestCase
             $processorBagConfigBuilder->getArgument(0)
         );
         self::assertEquals(
-            ['' => [0 => [['processor1', []]]]],
+            [
+                '' => [
+                    0 => [
+                        ['processor1.decorator', []],
+                        ['processor2.decorator2', []],
+                        ['processor3.decorator1', []]
+                    ]
+                ]
+            ],
             $processorBagConfigBuilder->getArgument(1)
         );
-        self::assertTrue($processor1->isPublic());
     }
 }